requirements.txt
PyYAML>=6.0
Jinja2>=3.1.0
rich>=13.0.0
textual>=0.40.0
jsonschema>=4.17.0
asteval>=0.9.31
pyperclip>=1.8.2
google-generativeai>=0.3.0
google-genai>=1.62.0
openai>=1.0.0
anthropic>=0.30.0
psutil>=5.9.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
pytest
pytest-asyncio
